Сделал поле base protected, В КПК с параметрами вызвал КБК

+++ src/Circle.java
@@ -19,7 +19,7 @@ public class Circle extends Figure
         System.out.print("Введите Радиус:");
         radius = scanner.nextInt();
     }
-    public float area()
+    public double area()
     {
         return PI * radius * radius;
     }
