+++ src/Figure.java
@@ -2,6 +2,6 @@ abstract public class Figure
 {
     public abstract  void set();
     public abstract void print();
-    public abstract float area();
+    public abstract double area();
 
 }
